Working GS comms

+++ Core/Inc/main.h
@@ -57,7 +57,7 @@ void handle_rf_rx_packet(uint8_t *Rx_buffer, size_t len);
 uint8_t get_rf_payload_len(uint8_t identifier);
 uint32_t Calculate_CRC32(CRC_HandleTypeDef *hcrc, uint8_t *payload_data, size_t len);
 void handle_payload_data(uint8_t identifier, uint8_t *payload_data);
-void send_rf_packet(uint8_t identifier, uint8_t* payload_data, size_t len);
+void send_rf_packet(uint16_t identifier, uint8_t* payload_data, size_t len);
 
 /* USER CODE END EFP */
